+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## [0.1.0] - 20/11/2025
 ### Added
+- ModifyProduct validation fields
+- RegisterProduct validation fields
+- Arrow to return register and modify product
+
+## [0.1.0] - 20/11/2025
+### Added
 - US for filter and order workshops.
 - Functional go back button to Supplies List Screen
 - Functional go back buttons to Orders and Inventory Screens